requirements.txt
alembic==0.9.8
arrow==0.12.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
pkg-resources==0.0.0
python-dateutil==2.7.0
python-dotenv==0.8.2
python-editor==1.0.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.5
SQLAlchemy-Utils==0.33.2
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
